refactor: rollback back `priority` value
Actually we need it. The images would only be loaded once closer to
viewport however if they take a while to download, user wouldn't
see the images immediately. Locally works fine but in PRD it's laggy.

`priority=true` makes sure they're all downloaded once the page loads.
There might be an initial hit but then scrolling and UX is smooth.